backend/requirements.txt
python-dotenv>=1.0.0
rich>=13.0.0
jupyter>=1.0.0
PyYAML>=6.0.0
SQLAlchemy>=2.0.0
Faker>=24.0.0
fastapi>=0.110.0
uvicorn>=0.29.0
langchain>=0.1.14
langgraph>=0.0.30
langchain-google-genai>=1.0.0
langchain-ollama>=0.1.0

psycopg2-binary==2.9.9

